monitor/requirements.txt
prometheus-client==0.6.0
pyeloqua==0.5.2
python-slugify==1.2.4
pytz==000.2
requests==2.18.3
python-dateutil==2.6.0
urllib3==1.22

# flask web page
flask==1.0.2
gunicorn==19.9.0
flask-httpauth==3.2.4